Adjusted everything in an attempt to make development easier
Added lumberjackGame.c and moved everything to it
Added menu to lumberjack so I can menu
