+++ main/CMakeLists.txt
@@ -21,6 +21,10 @@ idf_component_register(SRCS "swadge2024.c"
                             "modes/dance/dance.c"
                             "modes/dance/portableDance.c"
                             "modes/demo/demoMode.c"
+                            "modes/lumberjack/lumberjack.c"
+                            "modes/lumberjack/lumberjackGame.c"
+                            "modes/lumberjack/lumberjackEntity.c"
+                            "modes/lumberjack/lumberjackPlayer.c"
                             "modes/jukebox/jukebox.c"
                             "modes/mainMenu/mainMenu.c"
                             "modes/pong/pong.c"
@@ -65,6 +69,7 @@ idf_component_register(SRCS "swadge2024.c"
                                 "./modes/pong"
                                 "./modes/tunernome"
                                 "./modes/mainMenu"
+                                "./modes/lumberjack"
                                 "./modes/demo"
                                 "./modes/quickSettings")
